+++ layer.cpp
@@ -66,7 +66,7 @@ void Layer::add_mask(shared_ptr<Layer> mask)
 /*
  */
 /******************************************************************************/
-vector<unsigned int> Layer::get_bridges( shared_ptr<icoords> toolpath )
+vector<size_t> Layer::get_bridges( shared_ptr<icoords> toolpath )
 {
     return surface->get_bridges(dynamic_pointer_cast<Cutter>( manufacturer ), toolpath);
 }
