+++ layer.hpp
@@ -44,7 +44,7 @@ public:
 
     vector<shared_ptr<icoords> > get_toolpaths();
     shared_ptr<RoutingMill> get_manufacturer();
-    vector<unsigned int> get_bridges( shared_ptr<icoords> toolpath );
+    vector<size_t> get_bridges(shared_ptr<icoords> toolpath);
     string get_name()
     {
         return name;
